+++ doc/CHANGELOG.md
@@ -41,6 +41,7 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#2167` [Accessibility] Link form elements to their label - backlogs - new task
 * `#2258` [Accessibility] linearisation of issue show form
 * `#2607` [Agile] Display Storypoints in Taskboard view
+* `#3532` Fix: [API] It is possible to set statuses that are not allowed by the workflow
 * `#3649` A backlog item should get the highest position when moved to another backlog
 
 ## 3.0.5.pre1